# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
security:
  - bearerAuth: []
paths:
...
  /tasks:
    get:
      operationId: listTasks
      tags: [tasks]
      summary: List tasks
      description: >-
        Returns the authenticated user's tasks. Deleted tasks are excluded
        unless status=deleted is requested explicitly.
      parameters:
        - name: listId
          in: query
          required: false
          schema:
            type: string
            format: uuid
        - name: status
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/TaskStatus"
        - name: priority
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/PriorityLevel"
      responses:
        "200":
          description: The matching tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      operationId: createTask
      tags: [tasks]
      summary: Create a task
      description: >-
        The score is computed server-side from the task type and priority; a
        score supplied in the request body is ignored.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTaskRequest"
      responses:
        "201":
          description: The created task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
  /tasks/{taskId}:
    parameters:
      - $ref: "#/components/parameters/TaskId"
    get:
      operationId: getTask
      tags: [tasks]
      summary: Get a task
      responses:
        "200":
          description: The task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      operationId: updateTask
      tags: [tasks]
      summary: Update a task
      description: >-
        Changing taskTypeId or priority recomputes the score. Both are
        rejected on a completed task, whose score is a historical fact.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTaskRequest"
      responses:
        "200":
          description: The updated task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
    delete:
      operationId: deleteTask
      tags: [tasks]
      summary: Delete a task
      description: Soft delete - the row is retained so the action can be undone.
      responses:
        "200":
          description: The deleted task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /tasks/{taskId}/start:
    parameters:
      - $ref: "#/components/parameters/TaskId"
    post:
      operationId: startTask
      tags: [tasks]
      summary: Start a task
      responses:
        "200":
          $ref: "#/components/responses/TaskTransition"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
  /tasks/{taskId}/pause:
    parameters:
      - $ref: "#/components/parameters/TaskId"
    post:
      operationId: pauseTask
      tags: [tasks]
      summary: Pause a running task
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PauseTaskRequest"
      responses:
        "200":
          $ref: "#/components/responses/TaskTransition"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
  /tasks/{taskId}/resume:
    parameters:
      - $ref: "#/components/parameters/TaskId"
    post:
      operationId: resumeTask
      tags: [tasks]
      summary: Resume a paused task
      responses:
        "200":
          $ref: "#/components/responses/TaskTransition"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
  /tasks/{taskId}/stop:
    parameters:
      - $ref: "#/components/parameters/TaskId"
    post:
      operationId: stopTask
      tags: [tasks]
      summary: Stop a task without completing it
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StopTaskRequest"
      responses:
        "200":
          $ref: "#/components/responses/TaskTransition"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
  /tasks/{taskId}/complete:
    parameters:
      - $ref: "#/components/parameters/TaskId"
    post:
      operationId: completeTask
      tags: [tasks]
      summary: Complete a task
      description: >-
        Records completedAt and computes actualDurationMinutes with all
        paused time excluded.
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompleteTaskRequest"
      responses:
        "200":
          $ref: "#/components/responses/TaskTransition"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
  /tasks/{taskId}/restore:
    parameters:
      - $ref: "#/components/parameters/TaskId"
    post:
      operationId: restoreTask
      tags: [tasks]
      summary: Restore a deleted task
      description: The score is recomputed from the task's current type and priority.
      responses:
        "200":
          $ref: "#/components/responses/TaskTransition"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    TaskId:
      name: taskId
      in: path
      required: true
      schema:
        type: string
        format: uuid
  responses:
    TaskTransition:
      description: The task after the transition
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Task"
    Unauthorized:
      description: Missing or invalid credentials
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Problem"
    NotFound:
      description: No such resource for this user
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Problem"
    Conflict:
      description: The request conflicts with the resource's current state
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Problem"
    UnprocessableEntity:
      description: The request body failed validation
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Problem"
  schemas:
...
    Problem:
      type: object
      description: >-
        A machine-readable code plus copy that is safe to show the user.
        Per Principle III, no message assigns blame.
      properties:
        code:
          $ref: "#/components/schemas/ProblemCode"
        title:
          type: string
        detail:
          type: string
        fieldErrors:
          type: object
          description: Field name to message, for rendering inline beneath the field.
          additionalProperties:
            type: string
      required:
        - code
        - title
    ProblemCode:
      type: string
      enum:
        - Unauthorized
        - NotFound
        - DuplicateListName
        - InvalidTaskTransition
        - TaskAlreadyCompleted
        - UnknownTaskType
        - ValidationError
        - InternalError
    PriorityLevel:
      type: string
      enum: [low, medium, high, urgent]
    TaskStatus:
      type: string
      enum: [pending, in_progress, paused, completed, deleted]
    TaskType:
      type: object
      properties:
        id:
          type: string
          format: uuid
        type:
          type: string
        label:
          type: string
        score:
          type: integer
        isSystem:
          type: boolean
      required: [id, type, label, score, isSystem]
...
    Task:
      type: object
      properties:
        id:
          type: string
          format: uuid
        listId:
          type: string
          format: uuid
        taskTypeId:
          type: string
          format: uuid
        taskType:
          $ref: "#/components/schemas/TaskType"
        title:
          type: string
        description:
          type: string
        priority:
          $ref: "#/components/schemas/PriorityLevel"
        score:
          type: integer
          description: Server-computed snapshot; clients never send this.
        status:
          $ref: "#/components/schemas/TaskStatus"
        estimatedDurationMinutes:
          type: integer
        actualDurationMinutes:
          type: [integer, "null"]
        savedRemainingSeconds:
          type: [integer, "null"]
        totalPausedMs:
          type: integer
        startedAt:
          type: [string, "null"]
          format: date-time
        pausedAt:
          type: [string, "null"]
          format: date-time
        completedAt:
          type: [string, "null"]
          format: date-time
        deletedAt:
          type: [string, "null"]
          format: date-time
        dueDate:
          type: [string, "null"]
        position:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - listId
        - taskTypeId
        - taskType
        - title
        - description
        - priority
        - score
        - status
        - estimatedDurationMinutes
        - totalPausedMs
        - position
        - createdAt
        - updatedAt
...
    CreateTaskRequest:
      type: object
      properties:
        listId:
          type: string
          format: uuid
        taskTypeId:
          type: string
          format: uuid
        title:
          type: string
          minLength: 1
          maxLength: 100
        description:
          type: string
          maxLength: 500
        priority:
          $ref: "#/components/schemas/PriorityLevel"
        estimatedDurationMinutes:
          type: integer
          minimum: 1
          maximum: 480
        dueDate:
          type: [string, "null"]
        position:
          type: integer
      required: [listId, taskTypeId, title]
    UpdateTaskRequest:
      type: object
      properties:
        listId:
          type: string
          format: uuid
        taskTypeId:
          type: string
          format: uuid
        title:
          type: string
          minLength: 1
          maxLength: 100
        description:
          type: string
          maxLength: 500
        priority:
          $ref: "#/components/schemas/PriorityLevel"
        estimatedDurationMinutes:
          type: integer
          minimum: 1
          maximum: 480
        dueDate:
          type: [string, "null"]
        position:
          type: integer
    PauseTaskRequest:
      type: object
      properties:
        savedRemainingSeconds:
          type: integer
          minimum: 0
    StopTaskRequest:
      type: object
      properties:
        savedRemainingSeconds:
          type: integer
          minimum: 0
    CompleteTaskRequest:
      type: object
      properties:
        savedRemainingSeconds:
          type: integer
          minimum: 0
        totalPausedMs:
          type: integer
          minimum: 0
          description: >-
            Pause time accumulated on the device but not yet synced. Added to
            the stored total before the focused duration is computed.
...